# video-generation-platform/docs/openapi/video-api.yaml
openapi: 3.0.3
info:
  title: Dynamic Video Content Generation Platform API
  version: "1.0.0"
  description: |
    Professional API for dynamic video content generation with intelligent dual response system,
    AWS S3 storage, real-time progress tracking, and advanced orchestration capabilities.

    ## Features
    - **Dual Response System**: Quick processing (≤30 seconds) returns immediate results, longer jobs return async tracking
    - **Real-time Updates**: WebSocket and Server-Sent Events for live progress monitoring
    - **AWS S3 Integration**: Secure file storage with public URLs
    - **Advanced Orchestration**: Intelligent resource management and load balancing
    - **Multi-element Composition**: Support for up to 10 video/image elements per job

    ## Rate Limits
    - Upload endpoints: 10 requests per minute per IP
    - Status check endpoints: 100 requests per minute per IP
    - General API endpoints: 1000 requests per hour per IP

    ## Processing Limits
    - Maximum video input duration: 10 minutes
    - Maximum processing time: 10 minutes for async jobs
    - Maximum elements per job: 10
    - Supported formats: MP4, MOV, AVI
    - Maximum resolution: 7680x4320 (8K)
  contact:
    name: Video Generation Platform API Support
    email: api-support@videogeneration.platform
    url: https://docs.videogeneration.platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://videogeneration.platform/terms
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /video/result/{jobId}:
    get:
      summary: Get Job Status (Alternative Endpoint)
      description: Alternative endpoint for job status (same functionality as /videoresult/{jobId})
      tags:
        - Video Processing
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      description: Unique identifier for the video processing job
      schema:
        type: string
        pattern: "^job_[a-zA-Z0-9]+$"
        minLength: 10
        maxLength: 50
      example: "job_abc123def456"
...
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when the rate limit resets (Unix timestamp)
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: "RATE_LIMIT_EXCEEDED"
            message: "Rate limit exceeded. Try again in 60 seconds."
            statusCode: 429
            timestamp: "2024-01-15T10:30:00.000Z"
            correlationId: "req_123abc"
    JobNotFound:
      description: Job not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: "JOB_NOT_FOUND"
            message: "Video processing job not found"
            statusCode: 404
            timestamp: "2024-01-15T10:30:00.000Z"
            correlationId: "req_123abc"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred. Please try again later."
            statusCode: 500
            timestamp: "2024-01-15T10:30:00.000Z"
            correlationId: "req_123abc"
  schemas:
...
    JobStatusResponse:
      type: object
      required:
        - status
        - job_id
        - message
      properties:
        status:
          type: string
          enum: [processing, completed, failed, cancelled]
          description: Current job status
          example: "processing"
        job_id:
          type: string
          description: Unique job identifier
          example: "job_abc123def456"
        progress:
          type: string
          pattern: '^\d+%$'
          description: Processing progress percentage
          example: "65%"
        current_step:
          type: string
          description: Current processing step description
          example: "Encoding final video"
        message:
          type: string
          description: Status message
          example: "Video processing in progress"
        result_url:
          type: string
          format: uri
          description: Public URL to generated video (when completed)
          example: "https://cdn.videogeneration.platform/videos/abc123/output.mp4"
        file_size:
          type: string
          description: Size of generated video file (when completed)
          example: "15.2 MB"
        duration:
          type: string
          description: Duration of generated video (when completed)
          example: "00:02:45"
        processing_time:
          type: string
          description: Total processing time (when completed)
          example: "5m 23s"
        error:
          type: string
          description: Error message (when failed)
          example: "Invalid video codec in source file"
        estimated_time_remaining:
          type: string
          description: Estimated time until completion
          example: "2m 30s"
...
    ApiError:
      type: object
      required:
        - error
        - message
        - statusCode
        - timestamp
      properties:
        error:
          type: string
          description: Error code identifier
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Request validation failed"
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00.000Z"
        correlationId:
          type: string
          description: Request correlation ID for tracking
          example: "req_123abc"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...